requirements.txt
asgiref==3.5.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-filter==22.1
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-smalluuid==1.2.1
djangorestframework==3.14.0
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
oauth2client==4.1.3
oauthlib==3.2.0
phonenumbers==8.12.47
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.8
PyQRCode==1.2.1
pytz==2022.6
pyzbar==0.1.9
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
smalluuid==1.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
